requirements.txt
PyGithub==1.55
python-gitlab==2.8.0
PyYAML==5.4.1
